Move parse_macro_input to a different trait
