+++ tests/test_meta_item.rs
@@ -10,7 +10,6 @@
 
 extern crate proc_macro2;
 extern crate syn;
-#[macro_use]
 extern crate quote;
 
 use proc_macro2::{Ident, Literal, Span};
@@ -302,60 +301,6 @@ fn test_parse_nested_meta() {
     );
 
     assert_eq!(expected, syn::parse_str(raw).unwrap());
-}
-
-#[test]
-fn test_parse_quote_nested_meta() {
-    let actual: Vec<NestedMeta> = parse_quote!(5);
-
-    let expected = vec![
-        NestedMeta::Literal(lit(Literal::i32_unsuffixed(5)))
-    ];
-
-    assert_eq!(expected, actual);
-
-    let actual: Vec<NestedMeta> = parse_quote!(list(name2 = 6));
-
-    let expected = vec![
-        NestedMeta::Meta(
-            MetaList {
-                ident: ident("list").into(),
-                paren_token: Default::default(),
-                nested: punctuated![NestedMeta::Meta(
-                    MetaNameValue {
-                        ident: ident("name2").into(),
-                        eq_token: Default::default(),
-                        lit: lit(Literal::i32_unsuffixed(6)),
-                    }
-                    .into(),
-                )],
-            }
-            .into(),
-        )
-    ];
-
-    assert_eq!(expected, actual);
-
-    let actual: Vec<NestedMeta> = parse_quote!(5, list(name2 = 6));
-
-    let expected = vec![
-        NestedMeta::Literal(lit(Literal::i32_unsuffixed(5))),
-        NestedMeta::Meta(
-            MetaList {
-                ident: ident("list").into(),
-                paren_token: Default::default(),
-                nested: punctuated![NestedMeta::Meta(
-                    MetaNameValue {
-                        ident: ident("name2").into(),
-                        eq_token: Default::default(),
-                        lit: lit(Literal::i32_unsuffixed(6)),
-                    }
-                    .into(),
-                )],
-            }
-            .into(),
-        )
-    ];
 }
 
 fn run_test<T: Into<Meta>>(input: &str, expected: T) {
